+++ src/pages/AddCourse.jsx
@@ -29,6 +29,7 @@ import Typography from "@mui/material/Typography";
 import MuiInput from "@mui/material/Input";
 import {styled} from "@mui/material/styles";
 import PageHeader from "../components/common/PageHeader.jsx";
+import SubjectViewer from "../components/admin/SubjectsViewer.jsx";
 
 function EditCourse() {
     const navigate = useNavigate();
@@ -343,53 +344,8 @@ function EditCourse() {
                     </Form.Group>
 
                     {course.courseSubjects &&
-                        course.courseSubjects.map((subject, index) => (
-                            <Form.Group
-                                controlId={`formSubject${index}`}
-                                key={index}
-                                className="mt-3 mb-3"
-                            >
-                                <TextField
-                                    fullWidth
-                                    type="text"
-                                    label={`Subject #${index + 1}`}
-                                    id="outlined-size-normal"
-                                    defaultValue=""
-                                    value={subject}
-                                    onChange={(e) => handleSubjectsChange(e, index)}
-                                />
-                                {index >= 0 && (
-                                    <Button
-                                        variant="danger"
-                                        className="mt-2"
-                                        onClick={() =>
-                                            setCourse((prevCourse) => {
-                                                const updatedSubjects = [...prevCourse.courseSubjects];
-                                                updatedSubjects.splice(index, 1);
-                                                return {
-                                                    ...prevCourse,
-                                                    courseSubjects: updatedSubjects,
-                                                };
-                                            })
-                                        }
-                                    >
-                                        Remove This Subject
-                                    </Button>
-                                )}
-                            </Form.Group>
-                        ))}
-                    <Button
-                        variant="secondary"
-                        className="mt-2 mr-2 add-subject-btn"
-                        onClick={() =>
-                            setCourse((prevCourse) => ({
-                                ...prevCourse,
-                                courseSubjects: [...prevCourse.courseSubjects, ""],
-                            }))
-                        }
-                    >
-                        Add A Subject
-                    </Button>
+                        <SubjectViewer courseSubjects={course.courseSubjects} handleSubjectsChange={handleSubjectsChange} setCourse={setCourse}> </SubjectViewer>
+                    }
                 </Card>
                 <Button
                     variant="secondary"
