+++ src/pages/EditCourse.jsx
@@ -26,6 +26,7 @@ import Slider from "@mui/material/Slider";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import PageHeader from "../components/common/PageHeader.jsx";
+import SubjectViewer from "../components/admin/SubjectsViewer.jsx";
 
 function EditCourse() {
     const location = useLocation();
@@ -368,53 +369,8 @@ function EditCourse() {
                     </Form.Group>
 
                     {course.courseSubjects &&
-                        course.courseSubjects.map((subject, index) => (
-                            <Form.Group
-                                controlId={`formSubject${index}`}
-                                key={index}
-                                className="mt-3 mb-3"
-                            >
-                                <TextField
-                                    fullWidth
-                                    type="text"
-                                    label={`Subject #${index + 1}`}
-                                    id="outlined-size-normal"
-                                    defaultValue=""
-                                    value={subject}
-                                    onChange={(e) => handleSubjectsChange(e, index)}
-                                />
-                                {index >= 0 && (
-                                    <Button
-                                        variant="danger"
-                                        className="mt-2"
-                                        onClick={() =>
-                                            setCourse((prevCourse) => {
-                                                const updatedSubjects = [...prevCourse.courseSubjects];
-                                                updatedSubjects.splice(index, 1);
-                                                return {
-                                                    ...prevCourse,
-                                                    courseSubjects: updatedSubjects,
-                                                };
-                                            })
-                                        }
-                                    >
-                                        Remove This Subject
-                                    </Button>
-                                )}
-                            </Form.Group>
-                        ))}
-                    <Button
-                        variant="secondary"
-                        className="mt-2 mr-2 add-subject-btn"
-                        onClick={() =>
-                            setCourse((prevCourse) => ({
-                                ...prevCourse,
-                                courseSubjects: [...prevCourse.courseSubjects, ""],
-                            }))
-                        }
-                    >
-                        Add A Subject
-                    </Button>
+                        <SubjectViewer courseSubjects={course.courseSubjects} handleSubjectsChange={handleSubjectsChange} setCourse={setCourse}> </SubjectViewer>
+                    }
                 </Card>
                 <Button
                     variant="secondary"
